Fix string casing error with enum parsing of simple types.
Enum.Parse() and Enum.IsDefined are case sensitive, so trying to parse
e.g. int32 to SimpleType.Int32 does not work. We can't use lower case on
the enum because we hit reserved words in C# language. We also don't want
to make the comparison in case insensitive way because stRiNg is not a
type. So we make a compromise and char.ToUpper on the first character in
the string before matching with the enumeration.