Exclude immutable class boilerplate from Coverage
Implemented ExcludeFromCodeCoverage attribute with justification string
until the existing in System.Diagnostics is changed from internal to
public accessibility.

Updated coverage script to fix one bug and exclude / hide classes and
methods decorated with this attribute from code coverage reports.

Need to go back and parameterize the script if I stick with it. But maybe
we move off it once there is tooling change. TBD.